+++ shared/src/main/scala/citewriter/CiteWriter.scala
@@ -22,6 +22,7 @@ import scala.scalajs.js.annotation._
 
   def writeCitePropertyDef(opd:CitePropertyDef):String
   def writeCiteObject(co:CiteObject, cd:CiteCollectionDef):String
+  def writeCitePropertyValue(pi:CitePropertyImplementation):String
   def writeCitePropertyValue(pv:CitePropertyValue):String
   def writeCiteCollectionDef(ccd:CiteCollectionDef):String